Rename: 검색 페이지 클라이언트 컴포넌트 구조 수정 - 불필요한 props drilling을 보완하고자 container 컴포넌트를   하나의 컴포넌트로 합침 - container 1 & presentation 3 구조로 구현
