--- src/app/search/_components/container/BookmarkContainer.tsx
@@ -1,55 +1,0 @@
-"use client";
-
-import { Book } from "@/types/common";
-import BookmarkUI from "../presentation/BookmarkUI";
-import { useSearchStore } from "@/stores/searchStore";
-import { useLikedBookStore } from "@/stores/likedBooks";
-
-interface ContainerProps {
-  book: Book;
-}
-
-export default function BookmarkContainer({ book }: ContainerProps) {
-  const { openDropDownId, setOpenDropDownId } = useSearchStore();
-
-  // 전역 서재 상태
-  const { toRead, reading, done } = useLikedBookStore();
-  const combined = [...toRead, ...reading, ...done];
-
-  function checkDuplicate(isbn: string): boolean {
-    return combined.some((b) => b.isbn === isbn);
-  }
-
-  const isOpen = openDropDownId === book.isbn;
-
-  const handleDropDownToggle = () => {
-    if (openDropDownId === book.isbn) {
-      setOpenDropDownId(null);
-    } else {
-      setOpenDropDownId(book.isbn);
-    }
-  };
-
-  // 북마크 클릭 이벤트
-  const handleClickBookmark = (e: React.MouseEvent) => {
-    // 부모 article 클릭 이벤트 차단
-    e.stopPropagation();
-
-    // 중복 체크
-    if (checkDuplicate(book.isbn)) {
-      alert("이미 서재에 존재합니다!");
-      setOpenDropDownId(null);
-      return;
-    }
-
-    handleDropDownToggle();
-  };
-
-  return (
-    <BookmarkUI
-      book={book}
-      isOpen={isOpen}
-      onClickBookmark={handleClickBookmark}
-    />
-  );
-}
