+++ src/app/search/_components/presentation/Bookmark.tsx
@@ -1,27 +1,19 @@
-"use client";
-
 import Image from "next/image";
 import { Book } from "@/types/common";
-import DropDownContainer from "../container/DropDownContainer";
 import bookmarkIcon from "/public/icons/Bookmark.svg";
 
 interface BookmarkProps {
   book: Book;
-  isOpen: boolean;
-  onClickBookmark: (e: React.MouseEvent) => void;
+  onClickBookmark: (e: React.MouseEvent, isbn: string) => void;
 }
 
-export default function BookmarkUI({
-  book,
-  isOpen,
-  onClickBookmark,
-}: BookmarkProps) {
+export default function Bookmark({ book, onClickBookmark }: BookmarkProps) {
   return (
     <>
       <button
         type="button"
         className="group absolute right-5 top-4 cursor-pointer"
-        onClick={onClickBookmark}
+        onClick={(e) => onClickBookmark(e, book.isbn)}
       >
         <Image
           src={bookmarkIcon}
@@ -31,13 +23,6 @@ export default function BookmarkUI({
           className="transition-transform duration-100 ease-in group-hover:scale-110"
         />
       </button>
-      {isOpen && (
-        <DropDownContainer
-          title={book.title}
-          isbn={book.isbn}
-          thumbnail={book.thumbnail}
-        />
-      )}
     </>
   );
 }
