+++ src/app/search/_components/presentation/DropDown.tsx
@@ -1,14 +1,20 @@
 interface DropDownProps {
+  title: string;
+  isbn: string;
+  thumbnail: string;
   selectedBookshelf: string;
   handleSelectChange: (e: React.ChangeEvent<HTMLSelectElement>) => void;
-  handleConfirm: () => void;
+  handleConfirm: (title: string, isbn: string, thumbnail: string) => void;
   setOpenDropDownId: (isbn: string | null) => void;
 }
 
-export default function DropDownUI({
+export default function DropDown({
+  title,
+  isbn,
+  thumbnail,
   handleConfirm,
-  selectedBookshelf,
   handleSelectChange,
+  selectedBookshelf,
   setOpenDropDownId,
 }: DropDownProps) {
   return (
@@ -24,7 +30,7 @@ export default function DropDownUI({
         name="bookshelves"
         id="bookshelf-select"
         value={selectedBookshelf}
-        onChange={handleSelectChange}
+        onChange={(e) => handleSelectChange(e)}
         className="w-full border px-2 py-1 text-sm"
       >
         <option value="">-- 서재 선택 --</option>
@@ -34,7 +40,7 @@ export default function DropDownUI({
       </select>
       <div className="flex w-full items-center justify-between gap-2">
         <button
-          onClick={handleConfirm}
+          onClick={() => handleConfirm(title, isbn, thumbnail)}
           className="w-full rounded bg-font-textPrimary px-2 py-1 text-sm text-brown-1"
         >
           추가
